refs #10 Отдельные файлы для атрибутов
