+++ include/textum/aho_corasick.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
-#include <textum/trie.hpp>
+#include <textum/aho_corasick_state_attribute.hpp>
+#include <textum/basic_trie.hpp>
 #include <textum/type_traits/iterator_value.hpp>
 
 #include <cassert>
 #include <cstdint>
 #include <iterator>
-#include <limits>
 #include <queue>
 #include <type_traits>
 #include <unordered_map>
@@ -14,33 +14,6 @@
 
 namespace textum
 {
-    template <typename Integral>
-    constexpr auto not_set = std::numeric_limits<Integral>::max();
-
-    /*!
-        \brief
-            Атрибуты состояния автомата Ахо-Корасик
-
-        \details
-            Добавление экземпляра этой структуры к состоянию простого автомата позволяет
-            организовать автомат Ахо-Корасик.
-            Содержит суффиксную ссылку и принимаемую суффиксную ссылку.
-    */
-    template <typename Integral>
-    struct aho_corasick_state_attribute_t: trie_state_attribute_t
-    {
-        /*!
-            Суффиксная ссылка всегда есть. При этом она может соответствовать "пустому" суффиксу,
-            и в этом случае это будет ссылка на корень автомата.
-        */
-        Integral suffix_link = not_set<Integral>;
-        /*!
-            Это ссылка на наибольший принимаемый суффикс из существующих в автомате. Может быть
-            пустой, если такой суффикс отсутствует.
-        */
-        Integral accept_suffix_link = not_set<Integral>;
-    };
-
     /*!
         \brief
             Автомат Ахо-Корасик и соответствующие поисковые механизмы
