+++ include/textum/trie.hpp
@@ -1,30 +1,12 @@
 #pragma once
 
 #include <textum/basic_trie.hpp>
+#include <textum/trie_state_attribute.hpp>
 
 #include <cstdint>
 
 namespace textum
 {
-    /*!
-        \brief
-            Атрибуты состояния бора
-
-        \details
-            Добавление экземпляра этой структуры к состоянию простого автомата позволяет
-            организовать бор.
-            Содержит индикатор принимаемости.
-    */
-    struct trie_state_attribute_t
-    {
-        /*!
-            Это булево значение, которое сигнализирует о том, что данное состояние в автомате
-            является принимаемым, то есть существует последовательность, которая приводит автомат
-            из его корня в данное состояние.
-        */
-        bool is_accept = false;
-    };
-
     /*!
         \brief
             Бор (префиксное дерево) и соответствующие поисковые механизмы
