+++ include/textum/trie_state_attribute.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+namespace textum
+{
+    /*!
+        \brief
+            Атрибуты состояния бора
+
+        \details
+            Добавление экземпляра этой структуры к состоянию простого автомата позволяет
+            организовать бор.
+            Содержит индикатор принимаемости.
+    */
+    struct trie_state_attribute_t
+    {
+        /*!
+            Это булево значение, которое сигнализирует о том, что данное состояние в автомате
+            является принимаемым, то есть существует последовательность, которая приводит автомат
+            из его корня в данное состояние.
+        */
+        bool is_accept = false;
+    };
+} // namespace textum
